+++ src/structures/attributs/index.ts
@@ -1,3 +1,0 @@
-// ? Exporting every attributs
-
-export * from "./client";