+++ src/structures/methods/index.ts
@@ -1,4 +1,0 @@
-// ? Exporting every methods.
-
-export * from "./console";
-export * from "./variables";