+++ src/types/enums.ts
@@ -1,0 +1,6 @@
+export enum ColorCodes {
+    RED = 31,
+    GREEN = 32,
+    YELLOW = 33,
+    BLUE = 34
+}