+++ src/types/types.ts
@@ -1,12 +1,1 @@
-export type OS =
-    | "type"
-    | "distribution"
-    | "version"
-    | "uptime"
-    | "totalMemory"
-    | "freeMemory"
-    | "cpuModel"
-    | "cpuSpeed"
-    | "architecture"
-    | "platform";
-export type Process = "ping" | "nodeVersion" | "npmVersion" | "tevVersion";
+export type ColorIndexes = 31 | 32 | 33 | 34;